- Fix breakage that slipped in through merges - Fix JSLint complaints - Move gross psuedo-unit-tests from Async.js into a new Async-test.js (still not factored cleanly, though) - Fix some docs
